Add new Service and fix entities

+++ service/src/main/java/com/project/api/rest/service/repository/ToDoRepository.java
@@ -1,9 +1,7 @@
 package com.project.api.rest.service.repository;
 
 import com.project.api.rest.model.ToDo;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ToDoRepository extends CrudRepository<ToDo,Long> {
-
+public interface ToDoRepository extends CrudRepository<ToDo, Integer> {
 }
